Simplen v1.1.2
* Bug fixes from previous versions

+++ app.js
@@ -93,7 +93,7 @@ function copyLink() {
 	}
 }
 
-if(location.hash == "#about" || localStorage.getItem("visitedfirst") != "yes") {
+if(location.hash == "#about" || (localStorage.getItem("visitedfirst") != "yes" && location.hash.replace('#', '') == '')) {
 	location.hash = "PGgzPldoYXQncyBTaW1wbGVuPzwvaDM+CjxwPldpdGggPGI+U2ltcGxlbjwvYj4geW91IGNhbiBjcmVhdGUgeW91ciBvd24gd2VicGFnZSwgd2l0aCBhIHVuaXF1ZSBsaW5rLiBMYXRlciBvbiwgeW91IGNhbiDwn5SXc2hhcmUgeW91ciBjcmVhdGlvbiB3aXRoIG90aGVycywgZXZlbiB0aHJvdWdoIGEgcXIgY29kZS48L3A+CjxwPjxiPvCfkqEgSWRlYXM8L2I+Ojxicj4K8J+SrCBFeHByZXNzIHlvdXJzZWxmIHdpdGggbG9uZ2VyIGRlc2NyaXB0aW9ucyBhbmQgcG9zdHM8YnI+CvCfjqggTWFrZSBjYW52YXMgb3IgYXNjaWkgY3JlYXRpb25zPGJyPgrwn5K7IENyZWF0ZSB5b3VyIG93biBhcHAKPC9wPgo8YSBocmVmPSIjIiB0YXJnZXQ9Il9ibGFuayI+PGJ1dHRvbiBzdHlsZT0nY29sb3I6ICNmZmY7IGJhY2tncm91bmQ6ICM2YmEzZmY7IGJvcmRlci1yYWRpdXM6IDEwMHB4OyBwYWRkaW5nOiA4cHggMTZweDsgYm9yZGVyOiBub25lOyBjdXJzb3I6IHBvaW50ZXI7Jz5TdGFydCBjcmVhdGluZzwvYnV0dG9uPjwvYT4=";
 	localStorage.setItem("visitedfirst", "yes");
 }
